Add docker-compose file for dev
Also add a simple script for bootstrapping a new server with the
necessary apt packages for running the site

The README should tell you how to use docker-compose to do local
development. Next steps are to dockerize the site and all of the other
components we want running on a server.

Closes #64, I need to add followup tasks for SSM since this PR doesn't
cover that.